Workaround a missing giterr in revparse
In case of @{-N} revspec (which is an Nth previously checked out
branch or commit), libgit does not set a giterr and only return
GIT_ENOTFOUND and a NULL obj. In such cases, EGIT_CHECK_ERROR does not
return (it requires a non-NULL giterr_last()), and we crash on NULL
pointer (obj) dereference.

This patch adds a check for this situation, and a emulates what libgit
was supposed to do. Another patch is expected to libgit, so it should
work better in future releases.